+++ terra_geocrud/tests/test_models.py
@@ -1,16 +1,16 @@
 from tempfile import TemporaryDirectory
 
 from django.contrib.gis.geos import Point
-from django.core.exceptions import ValidationError
 from django.test import override_settings
 from django.test.testcases import TestCase
 from terra_geocrud.tests.factories import CrudViewFactory, FeaturePictureFactory
 
 from geostore.models import Feature
 from terra_geocrud.models import AttachmentCategory, AttachmentMixin, \
-    feature_attachment_directory_path, feature_picture_directory_path
+    feature_attachment_directory_path, feature_picture_directory_path, CrudViewProperty
 from terra_geocrud.tests import factories
 from .. import models
+from ..properties.schema import sync_layer_schema, sync_ui_schema
 
 
 class CrudModelMixinTestCase(TestCase):
@@ -50,30 +50,30 @@ class CrudViewTestCase(TestCase):
             }
         )
 
-    def test_clean_default_list_properties(self):
-        with self.assertRaises(ValidationError):
-            self.crud_view.default_list_properties.append('toto')
-            self.crud_view.clean()
-
-    def test_clean_feature_title_property(self):
-        with self.assertRaises(ValidationError):
-            self.crud_view.feature_title_property = 'toto'
-            self.crud_view.clean()
-
 
 @override_settings(MEDIA_ROOT=TemporaryDirectory().name)
 class FeaturePropertyDisplayGroupTestCase(TestCase):
     def setUp(self) -> None:
         self.crud_view = factories.CrudViewFactory()
-        self.group_1 = models.FeaturePropertyDisplayGroup.objects.create(crud_view=self.crud_view, label='test',
-                                                                         properties=['age'])
-        self.group_2 = models.FeaturePropertyDisplayGroup.objects.create(crud_view=self.crud_view, label='test2',
-                                                                         properties=['name'])
+        self.group_1 = models.FeaturePropertyDisplayGroup.objects.create(crud_view=self.crud_view, label='test')
+        self.group_2 = models.FeaturePropertyDisplayGroup.objects.create(crud_view=self.crud_view, label='test2')
+        CrudViewProperty.objects.create(view=self.crud_view, key="name",
+                                        group=self.group_2,
+                                        required=True,
+                                        json_schema={'type': "string", "title": "Name"},
+                                        ui_schema={'ui:widget': 'textarea'})
+        CrudViewProperty.objects.create(view=self.crud_view, key="age",
+                                        group=self.group_1,
+                                        json_schema={'type': "integer", "title": "Age"})
+        CrudViewProperty.objects.create(view=self.crud_view, key="country",
+                                        json_schema={'type': "string", "title": "Country"})
         self.feature = Feature.objects.create(geom=Point(0, 0, srid=4326),
                                               properties={"age": 10, "name": "jec", "country": "slovenija"},
                                               layer=self.crud_view.layer)
         self.template = factories.TemplateDocxFactory()
         self.crud_view.templates.add(self.template)
+        sync_layer_schema(self.crud_view)
+        sync_ui_schema(self.crud_view)
 
     def test_str(self):
         self.assertEqual(str(self.group_1), self.group_1.label)
@@ -84,8 +84,7 @@ class FeaturePropertyDisplayGroupTestCase(TestCase):
                               'required': [],
                               'title': 'test',
                               'type': 'object'
-                              }
-                             )
+                              })
         self.assertDictEqual(self.group_2.form_schema,
                              {'properties': {'name': {'title': 'Name', 'type': 'string'}},
                               'required': ['name'],
@@ -94,7 +93,6 @@ class FeaturePropertyDisplayGroupTestCase(TestCase):
                               })
         self.maxDiff = None
         self.assertDictEqual(self.crud_view.grouped_form_schema, {
-            "type": "object",
             "required": [],
             "properties": {
                 'country': {
@@ -115,11 +113,6 @@ class FeaturePropertyDisplayGroupTestCase(TestCase):
                 },
             }
         })
-
-    def test_clean(self):
-        with self.assertRaises(ValidationError):
-            self.group_1.properties.append('toto')
-            self.group_1.clean()
 
 
 class AttachmentCategoryTestCase(TestCase):